Add motor characterization script v2 for telemetry-based testing
- Implemented a comprehensive motor characterization tool that connects to an ESP32 via WiFi.
- The script enables telemetry for motor performance analysis, including latency, time constant (tau), deadzone, and saturation.
- Introduced step response tests for both left and right motors with detailed logging of telemetry data.
- Added sweep tests to evaluate deadzone and saturation characteristics across a range of command inputs.
- Results are saved in a structured format, including summary statistics and visual plots for analysis.

+++ ESP32/include/motor_drivers/speed_estimator.h
@@ -39,11 +39,11 @@ public:
   void reset();
   // Update estimator with current cumulative encoder count and timestamp
   // `now_us` in microseconds. Returns the new speed estimate (counts/sec).
-  float update(int64_t curCount, uint32_t now_us);
+  float update(int64_t curCount, uint64_t now_us);
   float get() const { return m_speed; }
 
 private:
-  uint32_t m_last_ts_us;
+  uint64_t m_last_ts_us;
   int64_t m_last_count;
   float m_speed;
   float m_alpha;
